tiny format change in parse.c
The guts of it is here. When using sendto_match_butone() which is =only=
used to do the "oper magic" messages. i.e. /msg $server.mask and /msg #hostmask
The prefix sent to remote servers is only the nick, not the nick!user@host.
The prefix sent to local users is nick!user@host