Lee's patch..here is the description he gave:
The following patch adds back support for bouncing modes from clients we
deop on an SJOIN.

Basically - when we bounce a clients status (either halfop or chanop) we
add a user to a chptr->deopped LL.  Whenever we receive a mode, we search
this list for the client, ignoring it if theyre found.

Whenever we change a clients membership on a channel (from anything, to
anything), we scan the deopped list, removing them if theyre there.