HOWTO-CVS: Correct a minor typo.
Makefile.in: We use a shell conditional to check whether include/setup.h
is available, and run configure if it isn't - better to take advantage
of make. build: now depends on include/setup.h, which in turn calls
configure.

configure.ac, configure: don't use -lresolv and -lnsl if function is
natively available.