This commit removes multiple channel member lists like chptr->chanops, chptr->chanops_voiced etc. This was originally designed as an optimalization, but indeed it introduced a whole layer of complexity. We came to a conclusion that the best approach would be having two lists: chptr->members and chptr->locmembers (the latter not supported yet, but we are about to add it).
This can speed up things as well, because now we have struct Membership,
which is used not only with chptr->members, but also with
client_p->user->channel. Therefore it allows us to determine if someone
is an op by checking only 10 entries for channels s/he is on (instead of
a channel ops list which may contain hundreds of users). Another advantage
is that it's now easy to modularise the code (in opposite to situation
when we have 12 members hard-coded lists).
Most of these code changes by Michael, reviewed, supplemented and
corrected by me.

Other modifications:
- fixed CLEARCHAN. It's wrong to send KICK messages to an oper which has
  not yet been informed that they're on the channel... Also changed
  prefix format from :n to :n!u@h when sending KICKs to local users.
- moved JOIN to core modules. Server, which doesn't understand the
  ":%s JOIN 0" message, could easily get desynced.
- miscellaneous cleanups and style corrections.